backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0  # Add this
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.5  # For PostgreSQL support
httpx>=0.24.0
pydantic-settings>=2.0.0